feat(jsonschema): Add option to preserve proto field names
Adds a `--preserving-proto-field-name` flag to the command-line tool. When this flag is used, the generated JSON schema will use the original protobuf field names instead of the default camelCase JSON names.

This provides more flexibility for users who need to generate schemas that match the output of tools like `google.protobuf.json_format.MessageToDict` when `preserving_proto_field_name` is set to `True`.