feat(jsonschema): Initial JSON schema support (#2)
This adds support for converting an XSD schema to jsonschema via the protobuf definition.

The reason for this is that although JSON isn't efficient as a storage format, it's useful to be able to describe the representation that LLMs will receive to them. The aim of converting via the protobuf definition is to ensure that the jsonschema matches what will result when converting the protobuf records to json.

Critically, comments that exist in the xsd are propagated to jsonschema, so that the documentation is preserved and made available.