feat: Add definitions_only option to JSON schema generator
This commit adds a `definitions_only` option to the JSON schema
generator. When this option is true, the generated schema will only
contain the `definitions` section, without the top-level `$ref`.
This is useful when the generated schema is intended to be used as
a library of definitions for other schemas.